Remove useUdp from the code
(cherry picked from commit d516589ecef868eeb8ebbed0cca94e246d6667bc)
